test fixes and test for vpc
