+++ otcextensions/tests/unit/osclient/vpc/v1/test_peering.py
@@ -15,8 +15,8 @@ from unittest.mock import call
 
 from osc_lib import exceptions
 
-from otcextensions.osclient.vpc.v2 import peering
-from otcextensions.tests.unit.osclient.vpc.v2 import fakes
+from otcextensions.osclient.vpc.v1 import peering
+from otcextensions.tests.unit.osclient.vpc.v1 import fakes
 
 from openstackclient.tests.unit import utils as tests_utils
